Clean up customer_id (billing), cost and change link between account_invoice* and fsm_order
