+++ fieldservice_account/__init__.py
@@ -1,5 +1,4 @@
 # Copyright (C) 2018 Open Source Integrators
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 from . import models
-from . import wizard
 from .init_hook import pre_init_hook
