+++ fieldservice_account/readme/CONTRIBUTORS.rst
@@ -1,2 +1,3 @@
 * Michael Allen <mallen@opensourceintegrators.com>
 * Serpent Consulting Services Pvt. Ltd. <support@serpentcs.com>
+* Raphaël Reverdy <raphael.reverdy@akretion.com>
