+++ fieldservice_account/readme/DESCRIPTION.rst
@@ -1,3 +1,1 @@
 This module adds the ability to link field service orders to invoices.
-It also adds the option to track the billing partner for field service
-locations and orders.
