--- fieldservice_account/views/account_invoice_view.xml
@@ -1,46 +1,0 @@
-<odoo>
-
-    <record id="invoice_form_fsm_contractor" model="ir.ui.view">
-        <field name="name">account.invoice.form.fsm.contractor</field>
-        <field name="model">account.invoice</field>
-        <field name="inherit_id" ref="account.invoice_supplier_form"/>
-        <field name="arch" type="xml">
-            <field name="date_due" position='after'>
-                <field name="fsm_order_id" readonly="1"/>
-            </field>
-        </field>
-    </record>
-
-    <record id="invoice_form_fsm_customer" model="ir.ui.view">
-        <field name="name">account.invoice.form.fsm.customer</field>
-        <field name="model">account.invoice</field>
-        <field name="inherit_id" ref="account.invoice_form"/>
-        <field name="arch" type="xml">
-            <field name="date_due" position='after'>
-                <field name="fsm_order_id" readonly="1"/>
-            </field>
-        </field>
-    </record>
-
-    <record id="view_account_invoice_fsm_tree" model="ir.ui.view">
-        <field name="name">view.account.invoice.fsm.tree</field>
-        <field name="model">account.invoice</field>
-        <field name="inherit_id" ref="account.invoice_supplier_tree"/>
-        <field name="arch" type="xml">
-            <field name="number" position="after">
-                <field name="fsm_order_id"/>
-            </field>
-        </field>
-    </record>
-
-    <record id="view_account_invoice_fsm_search" model="ir.ui.view">
-        <field name="name">view.account.invoice.fsm.search</field>
-        <field name="model">account.invoice</field>
-        <field name="inherit_id" ref="account.view_account_invoice_filter"/>
-        <field name="arch" type="xml">
-            <field name="number" position="after">
-                <field name="fsm_order_id" string="FSM Order"/>
-            </field>
-        </field>
-    </record>
-</odoo>
