+++ fieldservice_account_analytic/readme/DESCRIPTION.rst
@@ -1,2 +1,4 @@
 This module adds the ability to track your analytic accounts on field
 service orders.
+It also adds the option to track the billing partner for field service
+locations and orders.
