+++ fieldservice_account_analytic/views/fsm_order.xml
@@ -1,6 +1,6 @@
 <odoo>
 
-    <record id="fsm_order_form_accounting" model="ir.ui.view">
+    <record id="fsm_order_form_analytic" model="ir.ui.view">
         <field name="name">fsm.order.form.account</field>
         <field name="model">fsm.order</field>
         <field name="inherit_id" ref="fieldservice.fsm_order_form"/>
