+++ fieldservice_account_analytic/views/fsm_person.xml
@@ -12,7 +12,7 @@
     </record>
 
     <!-- Field Service Location Person Form View-->
-    <record id="fsm_location_person_form_view_accounting" model="ir.ui.view">
+    <record id="fsm_location_person_form_view_analytic" model="ir.ui.view">
         <field name="name">fsm.location.person.form</field>
         <field name="model">fsm.location.person</field>
         <field name="inherit_id" ref="fieldservice.fsm_location_person_form_view"/>
